requirements.txt
fastapi==0.89.1
uvicorn==0.19.0
pydantic==1.10.2
multipart==0.2.4
python-multipart==0.0.5
python-dotenv==0.21.0
fastapi-pagination==0.10.0
SQLAlchemy==1.4.46
psycopg2-binary==2.9.5
SQLAlchemy-Utils==0.39.0
python-ulid==1.1.0
pydantic[email]==1.10.2
PyJWT==2.8.0
fastapi-pagination==0.10.0